+++ src/ParametersTest.php
@@ -62,7 +62,7 @@ final class ParametersTest extends StructuredFieldTest
      */
     public function it_fails_to_instantiate_with_an_item_containing_already_parameters(): void
     {
-        $this->expectException(SyntaxError::class);
+        $this->expectException(ForbiddenStateError::class);
 
         Parameters::fromAssociative([
             'foo' => Item::from(